fix(conversations): res.ok checks, SWR revalidation, double-rename guard, extract ConversationTitle
