+++ src/app/api/conversation/route.test.ts
@@ -1,6 +1,7 @@
 import { GET, POST } from "./route";
 import {
   createConversation,
+  getOrCreateActiveConversation,
   listConversations,
 } from "@/lib/conversations";
 import { NextRequest } from "next/server";
@@ -20,10 +21,12 @@ jest.mock("next/server", () => ({
 jest.mock("@/lib/conversations", () => ({
   listConversations: jest.fn(),
   createConversation: jest.fn(),
+  getOrCreateActiveConversation: jest.fn(),
 }));
 
 const mockedListConversations = jest.mocked(listConversations);
 const mockedCreateConversation = jest.mocked(createConversation);
+const mockedGetOrCreateActiveConversation = jest.mocked(getOrCreateActiveConversation);
 
 const createMockRequest = (url: string, options: RequestInit = {}) => {
   const parsedUrl = new URL(url);
@@ -66,6 +69,24 @@ function makeConversation(overrides: Partial<{
 describe("Conversation API Routes", () => {
   beforeEach(() => {
     jest.clearAllMocks();
+  });
+
+  describe("GET ?active=true", () => {
+    it("should return the active conversation for a user", async () => {
+      const conv = makeConversation({ id: "conv-active", userId: "test-user-123" });
+      mockedGetOrCreateActiveConversation.mockResolvedValue(conv);
+
+      const request = createMockRequest(
+        "http://localhost:3000/api/conversation?active=true&userId=test-user-123"
+      );
+      const response = await GET(request);
+      const data = await response.json();
+
+      expect(response.status).toBe(200);
+      expect(data).toEqual({ conversation: conv });
+      expect(mockedGetOrCreateActiveConversation).toHaveBeenCalledWith("test-user-123");
+      expect(mockedListConversations).not.toHaveBeenCalled();
+    });
   });
 
   describe("GET /api/conversation", () => {
